refactor: presentation 을 각 컨텍스트에 배치한다
