+++ backend/src/main/java/com/festago/auth/AuthInterceptor.java
@@ -1,4 +1,4 @@
-package com.festago.presentation.auth;
+package com.festago.auth;
 
 import com.festago.auth.application.AuthExtractor;
 import com.festago.auth.application.TokenExtractor;
