+++ backend/src/main/java/com/festago/auth/AuthenticateContext.java
@@ -1,4 +1,4 @@
-package com.festago.presentation.auth;
+package com.festago.auth;
 
 import com.festago.auth.domain.Role;
 import org.springframework.stereotype.Component;
