+++ backend/src/test/java/com/festago/support/TestAuthConfig.java
@@ -1,7 +1,7 @@
 package com.festago.support;
 
+import com.festago.auth.AuthenticateContext;
 import com.festago.auth.application.AuthExtractor;
-import com.festago.presentation.auth.AuthenticateContext;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
